extract useVoteResults
